test(base-data-service): collapse fetchInfiniteQuery tests into BaseDataService.test.ts
